Arreglado el error de numero en ConsoleFile

+++ ConsoleFile.java
@@ -12,7 +12,7 @@ public class ConsoleFile{
 	public static String[] arrayListToStrings(ArrayList<String> al)
 	{
 		String []out = new String[al.size()];
-		byte count = 0;
+		short count = 0;
 		for (String token : al) {
 			out[count++] = token;
 			//System.out.print(token);
